Renamed '/hooks/utils' directory into '/hooks/core'
