+++ components/Home/SelectedClients/Clients/index.js
@@ -2,7 +2,7 @@ import { forwardRef, useState } from "react";
 import styles from "../SelectedClients.module.css";
 import { CLIENTS } from "@/constants/clients";
 import ChangeText from "@/components/Global/ChangeText";
-import { useHoverCapable } from "@/hooks/utils/useHoverCapable";
+import { useHoverCapable } from "@/hooks/core/useHoverCapable";
 import Link from "next/link";
 
 const ClientItem = forwardRef(({ client }, ref) => {
