+++ hooks/core/useConsoleMessage.js
