+++ hooks/core/useHoverCapable.js
