+++ hooks/core/useSupportsHEVC.js
